requirements.txt
# Dependencias principales de Django
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Soporte para HTTPS y SSL
django-sslserver==0.22
django-extensions==3.2.3
Werkzeug==3.1.3
pyOpenSSL==24.1.0

# Base de datos
psycopg2-binary==2.9.9

# Procesamiento de datos
gpxpy==1.5.0
pandas==2.1.1
python-dateutil==2.8.2
pytz==2023.3.post1

# Seguridad
bcrypt==4.0.1

# Configuración y utilidades
python-dotenv==1.0.0
Pillow==10.0.1

# Exportación de datos
reportlab==4.1.0  # Para generar PDFs
python-pptx==1.0.1  # Para generar presentaciones
openpyxl==3.1.2  # Para generar archivos Excel
